gets launchbay price in line with other equipment (#5224)
# About the pull request

reduces price of launchbay to 200, getting it in line with fulton and
medevac. Aims to set ground for omnisentry price changes ( and possible
further utility additions to launchbay?)

# Explain why it's good for the game

every other eqipment costs around 200 points and there is not any reason
to ever have more then one launchbay and you can sell it after you use
all the omnisentryguns you want (but that is pain, it counts as weapon
and all FMs that you made with it installed need to be remade). Unlike
with fulton, medevac and rappel that have real use when you buy second
or third one.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
balance: launchbay price 400 -> 200
/:cl:

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>