Fixes CAS killing players with ammo crates instead of bullets (#5205)
# About the pull request

Fixes #3211 by making GAU-21 and Laser Cannon ammo pass the weapon to
its `cause_data` rather than the ammo crate/battery.

# Explain why it's good for the game

The dropship is meant to be firing the ammunition inside the crate, not
the crate itself. (Although that is fun to imagine)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed the death message from GAU-21 and Laser Cannon strikes saying
that the player was killed by the ammo crate.
/:cl: