Admin audio cue toggle. (#5185)
# About the pull request
Changes the toggle for ARES interface pings to also include prayers (and
renames it due to this)
Also moved it from logs category to sound, because it's for sounds...
Added SPAN_BIGNOTICE which combines bold & big classes to get rid of
"<b><big></big></b>"

# Explain why it's good for the game

Admin QOL, the sound can be overly intrusive and larger text gets the
point across.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Test to confirm it works as intended.

</details>


# Changelog
:cl:
code: Added a new span class that combines bold and big.
code: Tweaked the way prayer sends notifications to be more efficient.
admin: Moved the prayer notification sound to a toggle preference,
combined with ARES Interface notifications.
/:cl: