Fixes being able to vend infinite alcohol (#5305)
# About the pull request

Fixes #5291.
This was caused by the booze vendor returning a brand new list of
products every time something asked for them with
`get_listed_products()`, rather than making a single list when
initialised in `populate_product_list()` and returning that.

I also edited the `get_listed_products()` proc of the 'Experimental
Tools' vendor (fancy synth items), since it seemed to be the only one
not returning a global list.
*(I feel like those would all be better as static variables on each
vendor, but there's probably a good reason that they're not.)*

# Explain why it's good for the game

Infinite booze is... bad? 🤔

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed being able to vend infinite alcohol.
/:cl: