Fixes the Queen Eye being stuck as immature (#5299)
# About the pull request

Makes the 'Queen Eye' mob update its name when the Queen passes the
'Immature' stage, by adding a `COMSIG_MOB_REAL_NAME_CHANGED` signal.

This would be better if it was more similar to TG's
`COMSIG_ATOM_UPDATE_NAME`, but from looking into that it would take a
*big* refactor to implement `update_name()`/`update_appearance()`, so
this'll probably do for the time being.

# Explain why it's good for the game

Fixes the queen eye being stuck as immature unless the queen toggles it
on and off.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**


https://github.com/cmss13-devs/cmss13/assets/57483089/329b8e92-29b3-4629-bd94-c0096bea6508

**After:**


https://github.com/cmss13-devs/cmss13/assets/57483089/49746267-03f5-4e62-aa5b-6f36927cc3e9

</details>


# Changelog
:cl:
fix: Fixed the Queen Eye still showing as "Immature" after the Queen
ages.
/:cl: