Makes the designation of a tunnel display in chat when entered (#5315)
# About the pull request

Makes the designation of a tunnel display in chat when a player enters
it.

# Explain why it's good for the game

I've quite often gone through a tunnel to drop a capture off at the
hive, and then had no idea which tunnel to select in order to get back.
It is possible to examine a tunnel before you go through it to see its
name, but this just streamlines the whole process.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**

![before](https://github.com/cmss13-devs/cmss13/assets/57483089/de1a4a73-0d39-456a-a7fd-b0ddf401dc4f)

**After:**

![after](https://github.com/cmss13-devs/cmss13/assets/57483089/75d139eb-5ea2-41b1-a003-e42b52d08ecd)

</details>


# Changelog
:cl:
qol: Made the designation of a tunnel display in chat when a player
enters it.
/:cl: