Tutorial System: Retutorializing (#5030)
# About the pull request
Reopening of #4442

Adds a tutorial system to the game, accessible from the lobby screen.
The tutorial system is entirely isolated from the main game, allowing
players to get a curated experience to be taught the mechanics of SS13
or specific roles within CM. The tutorial system is fully capable of
supporting a theoretically infinite amount of players at once, each
getting their own instance.

See below video for an example of the in-dev "Marine - Basic" tutorial.

https://www.youtube.com/watch?v=aWEtd6EAZWk 

# Explain why it's good for the game
Teaching new players how to play the game has always been a tough bit
for us, so why not add a full-on tutorial system to get people into the
know?


# To-Do:
This list is alive and will change over time.
If you are interested in coding a tutorial, know that it's very easy and
that most of the heavy lifting's done for you! You can find a [tutorial
creation guide
here](https://hackmd.io/@mRAdleXgRfmKqh97O8ixSA/BJQsmO8kT), and you can
additionally contact me on discord at any time with questions.

Also, you can find an example tutorial
[here](https://github.com/cmss13-devs/cmss13/pull/4442/files#diff-843b2f84360b9b932dfc960027992f2b5117667962bfa8da14f9a35f0179a926).

Backend:
- [x] TGUI
- [x] Add finished tutorials to save files
- [x] Communicate to players with little playtime
- [x] Suppress combat logs and similar done in tutorials

SS13:
- [x] Basics
- [x] Intents

Marine:

- [x] Basics
- [x] Medical
- [ ] Weaponry
- [ ] Comtech - Basics
- [ ] Medic - Basics
- [ ] FTL - Basics
- [ ] Smartgunner - Basics
- [ ] Specialist - Demolitionist
- [ ] Specialist - Scout
- [ ] Specialist - Pyrotechnician
- [ ] Specialist - Grenadier
- [ ] Specialist - Sniper
- [ ] Squad Leader - Basics

Xenomorph:

- [ ] Basics
- [ ] Builder Caste - Basics

# Changelog
:cl:
add: Added a tutorial system for various roles (and just general
information), find it in the lobby screen.
/:cl:

---------

Co-authored-by: fira <loyauflorian@gmail.com>