Modify organization for vendors of the Rifleman and the Team Leader role. (#5187)
# About the pull request
added binocular to rifleman because he access to the higher tier of it
already.
added the advance tier to team leader because their is no reason for
rifleman to have it and not their leader.
added the section because it's a section other job have to store engi
item(SL and comtech)
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
better logic on what items are available for marines for binoculars
change.
the new section is on other job so it will probably make it better to
orient ourself in vendors.
the two new item because i think they aren't OP and are under-used items
that some marine might take if they can buy them.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: creating a new section called engineering supplies for the rifleman
and the team leader.
balance: added standard binocular to the rifleman vendor at a cost of 5.
balance: added Range Finder and Laser Designator binoculars to Team
leader at a cost of 10 and 20.
balance: added ES-11 Mobile Fuel Canister and ME3 hand welder to Team
leader and rifleman for 5 each.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>