Fixes Nuke failing to detonate (#5338)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Under very specific circumstances (seen 2x or more on live), the nuke
broken logic of checking eligbile mobs for explosion will cause it to
crash completely and fail to detonate. The problem is primarily the
`loc` check at top, which is not a sufficient safeguard.

Also moves the explosion delay so that people die when the nuke explode
and not before...


# Testing Photographs and Procedure
Spawn in nuke and testers, safety off, proccall explode


# Changelog
:cl:
fix: Fixed Nuke failing to detonate under specific circumstances. It may
involve a little rodent.
fix: Nuke explosion will now kill mobs on its actual explosion rather
than at the start of the cinematic sequence.
/:cl: