Fix runtimes and multitool bricking if a sentry_computer is deleted (#5345)
# About the pull request

This PR resolves some runtimes and the bricking of a multitool if it is
loaded with a deleted sentry. The weakref resolve in the multitool
wasn't checked if the resolve passed. Additionally, the sentries would
remain encrypted to a non-existent computer. If bricking an encrypted
sentry is intended, it needs to be handled more gracefully.

# Explain why it's good for the game

Fixes a runtime when using a multitool linked to a deleted
sentry_computer:

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/4db7c411-aff1-4b56-9cee-b2cec6e05519)

Fixes a runtime when using said multitool to unpair the sentry from a
deleted sentry_computer.

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/946f002d-2246-4632-a554-e2cfd6ad5a9d)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

1. Link tuner w/ a laptop
2. Delete laptop
3. Spawn a new laptop
4. Use tuner on new laptop
5. Runtime when trying to clear because the weakref resolve wasn't
checked.


https://github.com/cmss13-devs/cmss13/assets/76988376/bfe447ce-a6dc-4d88-b3e8-b4ee6efb1daf

</details>


# Changelog
:cl: Drathek
fix: Fix handling of sentry_computer deletion
/:cl: