Fixes xeno ERT distress beacon announcements (#5355)
# About the pull request

Fixes #5354 by flipping the result of the "Would you like to broadcast
the beacon launch?" prompt.
Basically the prompt was setting a variable as if `TRUE` == "Yes a
broadcast should be launched", when it was actually `TRUE` == "No a
broadcast shouldn't be launched".

(The parameters on the `activate()` proc are pretty weird with one being
positive and one negative, but I figured I'd just do the fix here rather
than trying to refactor anything.)

# Explain why it's good for the game

Less MOOC announcements needed about ""bugs"" :P.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**


https://github.com/cmss13-devs/cmss13/assets/57483089/12a8e585-e820-43c9-a9f4-22f250c3603a

**After:**


https://github.com/cmss13-devs/cmss13/assets/57483089/4bce0026-e7f9-4ac1-8eee-e7d2b6137921

</details>


# Changelog
:cl:
fix: Fixed custom xeno ERTs sending a distress beacon announcement when
it shouldn't.
/:cl: