Modifies and Re-Enables Shivas Snowball (#5113)
# About the pull request

This PR re-enables Shivas Snowball, placing it back in to the rotation
and allowing it to be played.

This PR also makes several modifications to the map to allow this and
fix the issues that resulted in its prior removal. Primarily in
relocating the South East LZ to the South West. Overall this results in
the eastern half of the map being xeno territory.

The comms towers are also relocated into more sane locations. 

# Explain why it's good for the game

Shivas Snowball is not a bad map, it just needed some changes to allow
it to function within the current game mechanics (mainly how comms
towers work). I hope these changes will return Shivas to a minimum level
of playability.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
mapadd: Shivas Snowball is back in rotation. 
maptweak: The south east landing zone on Shivas has been relocated to
the south west, the dig site in the south west has been similarly
relocated.
maptweak: Comms towers on Shivas have been placed in more consistent
locations near each other, instead of some of them being on the opposite
sides of the map from each other.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>