refactor part1 change define in coulours.dm from COULOUR to COLOR (#5361)
# About the pull request

So the idea is to standartize and remove as much as possible the use of
colour.
in this first PR i wanted to start with define using colour in them.

i also stole define in _math.dm because they add a bunch of define that
would have been in conflict. and i didn't really understand why some
color where define there...

can understand for blood and pipe but why math would need colors
defines??

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: for the file coulours.dm and other files change define in it
from COULOUR to COLOR
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>