Fixes an exploit with Praetorian Oppressor hook (#5375)
# About the pull request

The cooldown was not applied until after `throw_atom()` which sleeps was
operated on which allowed a brief but usable window to cast the ability
twice if we hit something. We now apply cooldown before we start
throwing atoms around so this is no longer possible.


![dreamseeker_2024-01-03_21-58-31](https://github.com/cmss13-devs/cmss13/assets/25027759/cb0b3662-4a1d-4fd5-8e38-6e41c8229f80)

![dreamseeker_2024-01-03_22-01-19](https://github.com/cmss13-devs/cmss13/assets/25027759/e2bf14b5-df0b-428b-be5f-e67a5956a794)

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed exploit in praetorian hook caused by an oversight.
/:cl: