Update missing bitfields for VV (#5369)
# About the pull request

This PR simply updates the bitfield defines used to parse bitfield
numbers in the variable viewer as text. The biggest reason for this PR
is to support vis_flags that I frequently need to fiddle with.

This isn't an exhaustive look at all bitfields we may have defined. Feel
free to request any others and I can also add them.

# Explain why it's good for the game

Fixes 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/a9715efb-40f9-4cf0-ab34-83d56c6272de)
And editing it is simply a number 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/1049398e-1841-4ad1-9d25-5c319b6e17e3)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/1438c778-7bd3-4b89-9352-0179431e6693)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/43e3d62c-d232-483d-aa99-71e08490d32f)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/6de59999-4754-4d2c-abd5-659425b6dd8e)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/e002e018-f76f-4e6e-9adf-9f878625c246)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/4a6776df-258a-4cd0-bab1-09843b2d61ac)

</details>


# Changelog
:cl: Drathek
code: Add missing bitfield definitions for variable viewer (VV)
/:cl: