Fixes the CLs office Posters and a Pinup Poster which was in the wrong category (#5386)
# About the pull request

Title says it all apart from the new plant and the ATM change in the CLs
office aswell, just a QOL change.

# Explain why it's good for the game

PROUDLY REMEMBER IO! Posters were a bug which was fixed, until a PR
reverted that change. I fixed it again.

Pinup poster being wrong is also fixed.

Added a cactus in the stead of a basic plant in the CLs office because
the sprite looks cool.

Changed the ATM Machine in the CLs office purely because the name was a
useless mapedit which annoyed me.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Spruced up the CLs office by adding a new plant, changing the stuck
posters to be random every round
fix: Fixed the ATM Machine in the CLs office having Weyland-Yutani
Automatic Teller Machine and not Wey-Yu. Same thing, useless mapedit.
fix: Fixes the PROUDLY REMEMBER IO! Posters in the CLs office (Again)
fix: Puts the 29th Poster which was a pinup into the pinup posters and
not the music ones.
/:cl:

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>