changes the runner pounce tip to be more accurate (#5378)
# About the pull request

Pounces are ineffective on marines who are laying down. -> You can dodge
pounces that aren't aimed directly at you by laying down.




# Explain why it's good for the game

the tip isn't actually true because runners CAN pounce marines that are
resting, but **only** if the pounce is a spriteclick.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
spellcheck: Changed a tip related to runner pounce to be more accurate
/:cl: