Tunnels are converted to forsaken on evac (#5394)
# About the pull request
the other half of #5340, fixes #4376 
Tunnels are now forsaken!
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Forsaken xenos should be able to use tunnels

# Testing Photographs and Procedure
<details>
<summary>Open me for testing</summary>


https://github.com/cmss13-devs/cmss13/assets/54422837/25f7ce72-0773-4bce-896a-800054d6b232



https://github.com/cmss13-devs/cmss13/assets/54422837/e4954940-f39e-4bda-9756-5f8da3dcc6a0



</details>


# Changelog
:cl:
fix: Tunnels are now converted to forsaken on evac.
/:cl: