Fixes stack of flags anchoring after placing one down (#5409)
# About the pull request

Currently, when you put down a flag, the stack of flags in your hand
becomes anchored instead of the placed down flag. Fixes #5145

# Explain why it's good for the game

The actual flag gets anchored, not the stack of flags in your hands.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/49321394/fb9092a5-9c2e-4d86-a591-e9f793a62a2e

</details>


# Changelog

:cl:
fix: Placing flags now anchors the actual flag, not the stack of flags
in your hand.
/:cl:

Co-authored-by: Nanu308 <59782240+Nanu308@users.noreply.github.com>