Steelpoint Donor Armour Update (#5252)
# About the pull request

Effects a sprite update to the M4X Armour (Donor Armour). Also changes
the item description to be more in-universe, and adds alt-camo options
for the armour based on the map type.

All credit to @Zenith for the sprite work. Camo sprites are my doing. 

Updates to donor gear are allowed based on my review of past PRs. 

# Explain why it's good for the game

Updating 2015 sprites to 2023 standards is good. 


# Testing Photographs and Procedure

![Screenshot 2023-12-20 12 43
35](https://github.com/cmss13-devs/cmss13/assets/6595389/a6326f14-ad85-41fe-a2f8-97b5e109b8a7)



# Changelog
:cl: Zenith, Steelpoint
add: Updates Steelpoint's donor armour, replacing it with a new sprite,
camouflage options for all maps and new item descriptions that are lore
friendly.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>