Uncluttering Queen UI P1 (#5350)
# About the pull request
Puts the banish/readmit/devolve into one category of "manage hive"

# Explain why it's good for the game

Queen ui is cluttered as hell and disheartening for new players this
aims to clear it up and put it into managable parts, this is the part
one of it.


# Testing Photographs and Procedure

![image](https://github.com/cmss13-devs/cmss13/assets/47158596/2de0fa68-36b8-43df-93ae-5ba9218913d5)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Moves "Banish, Re-admit and De-evolving" queen abilities into one
list, making the queen ui less clutterier
/:cl:

---------

Co-authored-by: InsaneRed <prodexter31@outlook.comm>
Co-authored-by: Birdtalon <birdtalon@gmail.com>