Adds three new zippo lighters in the code (#5412)
# About the pull request

Adds a black, blue and Wey-Yu Zippo Lighter to the code.

Adds the black and blue variant to the gear select menu next to the
normal zippo.

# Explain why it's good for the game

We have these sprites and they aren't used, I don't see what's wrong
with a bit of variety in zippos apart from the normal and the golden
one.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added three new Zippo Lighter sprites into the code.
add: Added black and blue variants to the gear pref menu.
fix: fixed the NV Zippo Lighter being black.
/:cl:

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>