Defines cleanup (#5439)
# About the pull request

Cleans up some duplicate, unused, and/or obsolete (due to native DM
implementation) defines. Each change is isolated in its own commit if
any are objectionable.

- `Clamp()` - functionally identical to and replaced by native
`clamp()`, introduced in 513:
https://www.byond.com/docs/ref/#/proc/clamp
- `Tan()` - functionally identical to and replaced by native `tan()`,
introduced in 513: https://www.byond.com/docs/ref/#/proc/tan
- `roundNearest()` - functionally identical to and replaced by native
2-argument `round(x, 1)`: https://www.byond.com/docs/ref/#/proc/round
- `IsInRange()` - never used and identical to `ISINRANGE()` define,
elected to keep the latter as it is in the DEFINES folder
- `CEILING(x, 1)` - when 1 is the second argument, replaced by the
functionally identical and simpler `Ceiling()` define, which itself can
be replaced by native `ceil()` when we move to 515:
https://www.byond.com/docs/ref/#/proc/ceil
- `n_ceil()` - functionally identical to and replaced by `Ceiling()`
- `FLOOR(x, 1)` - when 1 is the second argument, replaced by the
functionally identical and simpler `Floor()` define, which itself can be
replaced by native `floor()` when we move to 515:
https://www.byond.com/docs/ref/#/proc/floor
- `LERP()` - a version was implemented in projectile code, moved it out
to the DEFINES folder and switched to a slightly more robust version
from TGMC:
https://github.com/tgstation/TerraGov-Marine-Corps/blob/2da5c237640d73e3e66ad79e34861e9682f4609c/code/__DEFINES/maths.dm#L87-L90


<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Less duplication leads to greater maintainability.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
No player-facing changes.