Fixes dead marines being able to use the M56D (#5451)
# About the pull request

Fixes dead marines being able to man and fire the M56D.

Also adds a `stat` check to the M2C as well. It doesn't have the same
problem as the M56, but when dragging it onto your body a progress
indicator will flash up before quickly disappearing, so this just stops
it appearing in the first place.

# Explain why it's good for the game

Dead marines firing machine guns is spooky.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

**Before:**


https://github.com/cmss13-devs/cmss13/assets/57483089/e4e6fd55-ed71-493f-9f52-4c517f4eec61

**After:**


https://github.com/cmss13-devs/cmss13/assets/57483089/6459571f-02b8-457f-bb14-a6105c11ab19

</details>


# Changelog
:cl:
fix: Fixed dead marines being able to man and fire the M56D.
/:cl: