Fixes larvae sometimes spawning in space (#5455)
# About the pull request

Fixes larvae sometimes spawning in space on Fiorina. (And any other
space-based maps)

I think this is being caused by the `do_buried_larva_spawn()` proc
looking for any `/turf/open` subtypes in range, but not performing any
checks on them.

# Explain why it's good for the game

Occasionally at roundstart on Fiorina a larva will spawn outside of the
station's walls in space, and need an admin to teleport them back
inside. Although it *is* pretty funny when it happens, it's probably
best to fix it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixed larvae sometimes spawning in space on Fiorina.
/:cl: