refactor to reduce the number variable using colour in them instead of color. (#5472)
# About the pull request
another PR to reduce the number of time colour is being used for
variables.

did some change of comment to be to standard format too.

left with 54 instance of colour in 15 differents files. most that is
left have color var that could create conflit so i would rather do in a
seperate PR later.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: refactor to reduce the number variable using colour in them
instead of color.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>