TGUI upgrades to support react migration (#5435)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Round of upgrades and changes to support migration to React for tgui. By
doing more smaller changes I hope it's easier to review and test.

# Explain why it's good for the game
We want to migrate from infernojs to react.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: upgraded root tgui files to typescript to closer align with tg
refactor: upgraded typescript to 4.9.4
refactor: upgraded yarn to 3.3.1
refactor: upgraded a bunch of tgui tooling
/:cl: