New minimap update sound (#5429)
# About the pull request

Replaces the Morse Code sound for minimap updates with another data
stream sound.

# Explain why it's good for the game

sos-morse-code.ogg is used in a LOT of places. Having a unique "hey your
minimap was updated" sound is beneficial because it informs you that
something has changed that isn't some of the many already used updates
that people may not care about.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
soundadd: added unique sound for minimap updates
/:cl: