Fix m56d and m2c crushing, labeling, and damage retention (#5481)
# About the pull request

This PR addresses 5 things regarding the m56d and m2c:
1. Both now properly retain labeler labels (though technically the m56d
post can lose its label - gun takes priority)
2. The m56d now retains its damage when disassembled
3. The m56d's post now has an implementation for handle_charge_collision
so they can croosh not fully assembled guns
4. It appeared possible to duplicate the m56d when a charger crushed it
if it both was destroyed by the impact and the update_health resulted in
the gun being refunded; this is no longer possible since the charge
logic returns early if it is destroyed
5. Vehicles running over a m56d post now undeploys the m56d correctly if
it had a gun attached (it would downgrade to just a post)

# Explain why it's good for the game

Fixes #5477 but I don't think this *ever* worked. Rather the issue is
purely when the gun is in this state:

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/c72209b5-07da-46ac-adaf-c202bff4489e)

Additionally, if we're going to have a mechanic to weld to fix the gun;
you shouldn't be able to circumvent this by simply disassembling it...
However, since the destruction of the m56d didn't have a guaranteed
chance to destroy the gun, technically this results in a fully repaired
gun again if you can make another post for it. It would need to work
more like the m2c if we wanted it to remain in a broken state in this
situation.

# Testing Photographs and Procedure
https://youtu.be/rg1mLFjhsEY (though I forgot to record crowbar
disassembly for the m56d too)

# Changelog
:cl: Drathek
fix: The M56D and the M2C now retain their labels when disassembled
fix: The M56D can now be charger crushed when not fully assembled
fix: The M56D being run over by a vehicle now undeploys the gun
correctly (so it retains the gun if it had a gun)
fix: Fix an edge case that could duplicate the M56D when charger crushed
balance: The M56D now retains the health of the gun through all of its
different disassembled states
/:cl: